Create Earth Engine Script for forest change trajectory assessment
This Google Earth Engine Script is used to view the random sample points using the least cloudy images / data from Landsat 5, 7 and 8, Sentinel 2 and PALSAR 1 and 2. Images are loaded with the oldest images at the top. The images were reviewed visually and the sample points classified according to the classes described in Annex 1 of the article.